refactor(amazonq): move utils functions to dedicated file
